+++ lua/terrortown/menus/gamemode/equipment.lua
@@ -30,10 +30,6 @@ function CLGAMEMODEMENU:InitializeVirtualMenus()
     for i = 1, #equipments do
         local equipment = equipments[i]
 
-        if not equipment.isValidEquipment then
-            continue
-        end
-
         counter = counter + 1
 
         virtualSubmenus[counter] = tableCopy(equipmentMenuBase)