+++ CHANGELOG.md
@@ -47,6 +47,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - UI for grenade throw arcs from [colemclaren's TTT fork](https://github.com/colemclaren/ttt/blob/master/addons/moat_addons/lua/weapons/weapon_tttbasegrenade.lua#L293-L353) (integrated by @EntranceJew)
 - `gameEffects` library for global effects that are useful, such as starting fires (by @EntranceJew)
 - Added weapon pickup sounds when picking up weapons manually (by @TimGoll)
+- Added further missing German translation (by @NickCloudAT)
 
 ### Changed
 
@@ -89,10 +90,11 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Hidden when observing a player in first person view
 - Your own spectator nametag will not display when looking directly up in post-round (by @EntranceJew)
 - Made sure the last weapon is selected by default if the current weapon is removed; overwrite `OnRemove` to prevent that (by @TimGoll)
+- Changed the way weapon icon caching is working to make sure all weapons always have a cached icon material (by @TimGoll)
 
 ### Fixed
 
-- Fixed database now properly saving boolean `false` values
+- Fixed database now properly saving boolean `false` values (by @ZenBre4ker)
 - Fixed cached weapons not being selected after giving them back to the owner (by @TimGoll)
 - The roundendscreen can now be closed with the correct Binding (by @ZenBre4ker)
 - Fixed last seen player being wrongly visible for every search instead of only public policing role search (by @TimGoll)
@@ -106,7 +108,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed font initialization to not trip engine font fallback behavior (by @EntranceJew)
 - Fixed the decoy producing a wrong colored icon for other teams (by @NickCloudAT)
 - Fixed the scoreboard being stuck open sometimes if the inflictor was no weapon (by @TimGoll)
-- Fixed door health displaying as a humongous string of decimals
+- Fixed door health displaying as a humongous string of decimals (by @EntranceJew)
+- Fixed weapons that use the wrong weapon base from throwing errors in the F1 menu (by @TimGoll)
 
 ### Removed