+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Fixed
 
+## [v0.13.1b](https://github.com/TTT-2/TTT2/tree/v0.13.1b) (2024-02-27)
+
+### Fixed
+
 - Fixed rendering for weapons not based on the TTT2 base (e.g. TFA) (by @TimGoll)
 - Fixed bodysearch entries not having a title breaking the rendering (by @TimGoll)
 - Fixed `GetViewModel` error on the client when joining a server (by @TimGoll)